replace find head in cahe imp
