+++ packages/next/src/client/components/router-reducer/reducers/find-head-in-cache.test.tsx
@@ -95,10 +95,7 @@ describe('findHeadInCache', () => {
       ]),
     }
 
-    const result = findHeadInCache(
-      cache.parallelRoutes.get('children')!,
-      routerTree[1]
-    )
+    const result = findHeadInCache(cache, routerTree[1])
 
     expect(result).toMatchObject(
       <>
